Create native/window.h
