+++ src/window.c
@@ -1,12 +1,6 @@
 #include "window.h"
 
-// TODO find a better place for these!
-void native_window_raze(UTOX_WINDOW *win);
-UTOX_WINDOW *native_window_create_video(int x, int y, int w, int h);
-UTOX_WINDOW *native_window_find_notify(void *win);
-UTOX_WINDOW *native_window_create_notify(int x, int y, int w, int h, void *panel);
-void native_window_tween(UTOX_WINDOW *win);
-void native_window_create_screen_select(void);
+#include "native/window.h"
 
 void window_raze(UTOX_WINDOW *window) {
     native_window_raze(window);
@@ -20,7 +14,7 @@ UTOX_WINDOW *window_find_notify(void *window) {
     return native_window_find_notify(window);
 }
 
-UTOX_WINDOW *window_create_notify(int x, int y, int w, int h, void *panel) {
+UTOX_WINDOW *window_create_notify(int x, int y, int w, int h, PANEL *panel) {
     return native_window_create_notify(x, y, w, h, panel);
 }
