+++ src/window.h
@@ -4,18 +4,7 @@
 #include <stdbool.h>
 
 typedef struct native_window UTOX_WINDOW;
-/* The following is a hollow struct with window vars common across all platforms
- * Each platform's window struct starts with this, then follows with their own
- * specific includes. */
-struct utox_window {
-    UTOX_WINDOW *next;
-
-    int x, y;
-    unsigned w, h;
-
-    /* TODO do we want to include ui.c here to use PANEL here? */
-    void *panel;
-};
+typedef struct panel PANEL;
 
 /** window_raze()
  *
@@ -42,7 +31,7 @@ UTOX_WINDOW *window_find_notify(void *window);
  *
  * Create an interactive notification window and struct.
  */
-UTOX_WINDOW *window_create_notify(int x, int y, int w, int h, void *panel);
+UTOX_WINDOW *window_create_notify(int x, int y, int w, int h, PANEL *panel);
 
 /** window_tween()
  *
@@ -58,13 +47,5 @@ void window_tween(UTOX_WINDOW *win);
  * screen sharing or desktop images.
  */
 void window_create_screen_select(void);
-
-/**
- * Sets the target of the next series of drawing commands.
- *
- * Returns true if the window was changed.
- *         false if the window is the same.
- */
-bool native_window_set_target(UTOX_WINDOW *new_win);
 
 #endif
