+++ src/windows/window.h
@@ -1,9 +1,9 @@
 #ifndef WIN_WINDOW_H
 #define WIN_WINDOW_H
 
-#include "../window.h"
+#include "../native/window.h"
 
-#include <windows.h>
+#include <windef.h>
 
 struct native_window {
     struct utox_window _;
@@ -18,24 +18,7 @@ struct native_window {
     HBITMAP draw_BM;
 };
 
-UTOX_WINDOW main_window;
-
-HINSTANCE curr_instance;
-
-void native_window_init(HINSTANCE instance);
-
-void native_window_raze(UTOX_WINDOW *window);
-
-UTOX_WINDOW *native_window_create_main(int x, int y, int w, int h);
-
-HWND native_window_create_video(int x, int y, int w, int h);
-
-UTOX_WINDOW *native_window_create_notify(int x, int y, int w, int h, void *panel);
-
-UTOX_WINDOW *native_window_find_notify(void *window);
-
-void native_window_create_screen_select();
-
-void native_window_tween(UTOX_WINDOW *win);
+extern UTOX_WINDOW main_window;
+extern HINSTANCE curr_instance;
 
 #endif
