+++ src/xlib/window.h
@@ -3,6 +3,8 @@
 
 #include "../window.h"
 
+#include "../native/window.h"
+
 #include <X11/cursorfont.h>
 #include <X11/extensions/Xrender.h>
 #include <X11/extensions/XShm.h>
@@ -11,8 +13,8 @@
 #include <X11/Xlib.h>
 #include <X11/Xutil.h>
 
-#include <inttypes.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 Display *display;
 Screen  *default_screen;
@@ -48,20 +50,6 @@ struct native_window main_window;
 struct native_window popup_window;
 struct native_window scr_grab_window;
 
-bool native_window_init(void);
-
-UTOX_WINDOW *native_window_create_main(int x, int y, int w, int h, char **argv, int argc);
-
-UTOX_WINDOW *native_window_create_video(int x, int y, int w, int h);
-
-UTOX_WINDOW *native_window_find_notify(void *window);
-
-UTOX_WINDOW *native_window_create_notify(int x, int y, int w, int h, void *panel);
-
 void window_set_focus(UTOX_WINDOW *win);
-
-void native_window_tween(UTOX_WINDOW *win);
-
-void native_window_create_screen_select();
 
 #endif
